+++ src/plugin.hpp
@@ -33,6 +33,7 @@ extern Model* modelHolderCompact;
 extern Model* modelHolder8;
 extern Model* modelModulator;
 extern Model* modelModulator7;
+extern Model* modelModulator7Compact;
 extern Model* modelParking;
 extern Model* modelShifter;
 extern Model* modelSickoAmp;